+++ src/main/java/com/yooksi/betterarchery/init/Items.java
@@ -1,6 +1,5 @@
 package com.yooksi.betterarchery.init;
 
-import com.yooksi.betterarchery.common.BetterArchery;
 import com.yooksi.betterarchery.item.*;
 
 /** All custom items are initialized and stored here. */