+++ src/main/java/com/yooksi/betterarchery/item/RecurveBow.java
@@ -1,11 +1,12 @@
 package com.yooksi.betterarchery.item;
 
-import com.yooksi.betterarchery.common.BetterArchery;
-
 public class RecurveBow extends ArchersBow
 {
 	public RecurveBow()
 	{
 		pullingSpeedMult = 0.80F;
+		arrowSpeedMult = 1.25F;
+		
+		this.setMaxDamage(256);
 	}
 }