+++ src/main/java/com/yooksi/betterarchery/item/RegularBow.java
@@ -1,18 +1,5 @@
 package com.yooksi.betterarchery.item;
 
-import javax.annotation.Nullable;
-
-import com.yooksi.betterarchery.common.BetterArchery;
-
-import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.item.IItemPropertyGetter;
-import net.minecraft.item.ItemBow;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.world.World;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-
 public class RegularBow extends ArchersBow
 {